+++ tests/functional/visit-object.spec.ts
@@ -24,6 +24,18 @@ test.describe('visit object', () => {
 		});
 	});
 
+	test('has the next page html', async ({ page }) => {
+		await page.evaluate(() => {
+			window._swup.hooks.before('content:replace', (visit) => {
+				window.data = visit.to.html;
+			});
+		});
+		await navigateWithSwup(page, '/page-2.html');
+		await expectToBeAt(page, '/page-2.html', 'Page 2');
+		await page.waitForSelector('html:not(.is-changing)');
+		expect(await page.evaluate(() => window.data)).toMatch(/<h1>Page 2/i);
+	});
+
 	test('passes along click trigger and event', async ({ page }) => {
 		await page.evaluate(() => {
 			window.data = { el: null, event: null };
@@ -66,6 +78,41 @@ test.describe('visit object', () => {
 		await page.evaluate(() => {
 			window._swup.hooks.on('visit:start', (visit) => visit.animation.animate = false);
 		});
-		await expectSwupAnimationDuration(page, 0);
+		await expectSwupAnimationDuration(page, { out: 0, in: 0 });
+	});
+
+	test('allowings waiting for next page before animating', async ({ page }) => {
+		await page.evaluate(() => {
+			window._swup.hooks.on('visit:start', (visit) => {
+				visit.animation.wait = true;
+			});
+			window._swup.hooks.before('visit:transition', (visit) => {
+				window.data = visit.to.html;
+			});
+		});
+		await navigateWithSwup(page, '/page-2.html');
+		await expectToBeAt(page, '/page-2.html', 'Page 2');
+		expect(await page.evaluate(() => window.data)).toMatch(/<h1>Page 2/i);
+	});
+
+	test('marks native animations', async ({ page }) => {
+		const supported = await page.evaluate(() => !!document.startViewTransition);
+
+		// `visit.animation.native` should always be false when `native` option is false
+		await page.evaluate(() => {
+			window._swup.hooks.on('visit:start', (visit) => (window.data = visit.animation.native));
+		});
+		await navigateWithSwup(page, '/page-2.html');
+		await expectToBeAt(page, '/page-2.html', 'Page 2');
+		expect(await page.evaluate(() => window.data)).toBe(false);
+
+		// `visit.animation.native` should be true if `native` option is true and browser supports it
+		await page.goto('/animation-native.html');
+		await page.evaluate(() => {
+			window._swup.hooks.on('visit:start', (visit) => (window.data = visit.animation.native));
+		});
+		await navigateWithSwup(page, '/page-3.html');
+		await expectToBeAt(page, '/page-3.html', 'Page 3');
+		expect(await page.evaluate(() => window.data)).toBe(supported);
 	});
 });